reduce certain test sizes
I believe the previous solution was to use makeSmall in the Arbitrary
instances for certain things, so as to keep blocks etc. particularly
small. Better solution is here: in those tests which (for reasons
unknown) cause a developer's desktop to lock up when run with maximum
size 100, reduce the maximum size of the test.